Aggiunti gli ingredienti al DB con le relazioni ed aggiunta la visualizzazione nella Show e nei form Create/Edit

+++ Controllers/PizzasController.cs
@@ -1,4 +1,5 @@
 ﻿using la_mia_pizzeri_crud_mvc.Models;
+using la_mia_pizzeria_crud_mvc.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
@@ -34,10 +35,17 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
                 db.Add(new Category("Pizza di mare"));
                 db.SaveChanges();
 
-                db.Add(new Pizza("Margherita", "La pizza Margherita è buona", "https://www.finedininglovers.it/sites/g/files/xknfdk1106/files/styles/recipes_1200_800_fallback/public/fdl_content_import_it/margherita-50kalo.jpg?itok=v9nHxNMS", 10.50m, 1));
-                db.Add(new Pizza("Napoli", "La pizza Napoli è buona", "https://media-cdn.tripadvisor.com/media/photo-s/18/03/98/d6/received-665664433902722.jpg", 14.50m, 2));
-                db.Add(new Pizza("Romana", "La pizza Romana è buona", "https://recipesblob.oetker.com/files/95bdfe7334364b41b557c734cd1c64c4/889e39b112414a9aa2b3ae5a9f787f6b/1272x764/pizza-alla-romanajpg.jpg", 17.50m, 3));
-                db.Add(new Pizza("4 Gusti", "La pizza 4 Gusti è buona", "https://media-cdn.tripadvisor.com/media/photo-s/07/61/12/f1/pizza-4-gusti.jpg", 5.50m, 4));
+                db.Add(new Ingredient("Acciughe"));
+                db.Add(new Ingredient("Pomodoro"));
+                db.Add(new Ingredient("Wurstel"));
+                db.Add(new Ingredient("Funghi"));
+                db.Add(new Ingredient("Panna"));
+                db.SaveChanges();
+
+                db.Add(new Pizza("Margherita", "La pizza Margherita è buona", "https://www.finedininglovers.it/sites/g/files/xknfdk1106/files/styles/recipes_1200_800_fallback/public/fdl_content_import_it/margherita-50kalo.jpg?itok=v9nHxNMS", 10.50m, 1, new List<Ingredient>() { db.Ingredients.Find(1) , db.Ingredients.Find(2) }));
+                db.Add(new Pizza("Napoli", "La pizza Napoli è buona", "https://media-cdn.tripadvisor.com/media/photo-s/18/03/98/d6/received-665664433902722.jpg", 14.50m, 2, new List<Ingredient>() { db.Ingredients.Find(2), db.Ingredients.Find(3) }));
+                db.Add(new Pizza("Romana", "La pizza Romana è buona", "https://recipesblob.oetker.com/files/95bdfe7334364b41b557c734cd1c64c4/889e39b112414a9aa2b3ae5a9f787f6b/1272x764/pizza-alla-romanajpg.jpg", 17.50m, 3, new List<Ingredient>() { db.Ingredients.Find(4), db.Ingredients.Find(5) }));
+                db.Add(new Pizza("4 Gusti", "La pizza 4 Gusti è buona", "https://media-cdn.tripadvisor.com/media/photo-s/07/61/12/f1/pizza-4-gusti.jpg", 5.50m, 4, new List<Ingredient>() { db.Ingredients.Find(1), db.Ingredients.Find(5) }));
                 db.SaveChanges();
             }
 
@@ -51,17 +59,26 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
 
         public IActionResult Show(int id)
         {
-            Pizza pizza = db.Pizzas.Where(pizzas => pizzas.PizzaId == id).Include("Category").First<Pizza>();
+            Pizza pizza = db.Pizzas.Where(pizzas => pizzas.PizzaId == id).Include("Category").Include("Ingredients").FirstOrDefault();
 
-            return View(pizza);
+            if (pizza == null)
+            {
+                return NotFound($"La pizza con id {id} non è stata trovata");
+            }
+            else
+            {
+                return View(pizza);
+            }
         }
 
         [HttpGet]
         public IActionResult Create()
         {
-            PizzaWithCategory pizzaWithCategory = new();
-
-            pizzaWithCategory.Categories = db.Categories.ToList();
+            PizzaWithCategory pizzaWithCategory = new()
+            {
+                Categories = db.Categories.ToList(),
+                Ingredients = db.Ingredients.ToList()
+            };
 
             return View(pizzaWithCategory);
         }
@@ -71,11 +88,16 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
         public IActionResult Store(PizzaWithCategory model)
         {
             model.Categories = db.Categories.ToList();
+            model.Ingredients = db.Ingredients.ToList();
 
             if (!ModelState.IsValid)
             {
+                //model.Categories = db.Categories.ToList();
+                //model.Ingredients = db.Ingredients.ToList();
                 return View("Create", model);
             }
+
+            model.Pizza.Ingredients = db.Ingredients.Where(ingredient => model.SelectedIngredients.Contains(ingredient.IngredientId)).ToList<Ingredient>();
 
             db.Add(model.Pizza);
             db.SaveChanges();
@@ -86,7 +108,7 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
-            Pizza pizza = db.Pizzas.Find(id);
+            Pizza pizza = db.Pizzas.Include("Category").Include("Ingredients").Where(pizza => pizza.PizzaId == id).FirstOrDefault();
 
             if (pizza == null)
             {
@@ -97,8 +119,8 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
                 PizzaWithCategory pizzaWithCategory = new();
 
                 pizzaWithCategory.Pizza = pizza;
-
                 pizzaWithCategory.Categories = db.Categories.ToList();
+                pizzaWithCategory.Ingredients = db.Ingredients.ToList();
 
                 return View(pizzaWithCategory);
             }
@@ -110,19 +132,33 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
         public IActionResult Update(int id, PizzaWithCategory model)
         {
             model.Categories = db.Categories.ToList();
+            model.Ingredients = db.Ingredients.ToList();
 
             if (!ModelState.IsValid)
             {
+                //model.Categories = db.Categories.ToList();
+                //model.Ingredients = db.Ingredients.ToList();
                 return View("Edit", model);
             }
 
-            Pizza pizza = db.Pizzas.Find(id);
+            Pizza pizza = db.Pizzas.Include("Category").Include("Ingredients").Where(pizza => pizza.PizzaId == id).First();
 
             pizza.Name = model.Pizza.Name;
             pizza.Description = model.Pizza.Description;
             pizza.Image = model.Pizza.Image;
             pizza.Price = model.Pizza.Price;
             pizza.CategoryId = model.Pizza.CategoryId;
+
+            List<Ingredient> SelectedIngredients = db.Ingredients.Where(ingredient => model.SelectedIngredients.Contains(ingredient.IngredientId)).ToList();
+
+            foreach (Ingredient ingredient in SelectedIngredients)
+            {
+                if (!pizza.Ingredients.Contains(ingredient))
+                {
+                    pizza.Ingredients.Add(ingredient);
+                }
+                
+            }
 
             //db.Pizzas.Update(model);
             db.SaveChanges();
@@ -136,10 +172,19 @@ namespace la_mia_pizzeri_crud_mvc.Controllers
         {
             Pizza pizza = db.Pizzas.Find(id);
 
-            db.Pizzas.Remove(pizza);
-            db.SaveChanges();
+            if (pizza == null)
+            {
+                return NotFound();
+                
+            }
+            else
+            {
+                db.Pizzas.Remove(pizza);
+                db.SaveChanges();
 
-            return RedirectToAction("Index");
+                return RedirectToAction("Index");
+            }
+
         }
 
     }
